fixed bug

+++ esc/processer.go
@@ -9,7 +9,7 @@ import (
 )
 
 func Process(ctx context.Context, days, startHour uint32) *common.Activation {
-	service, err := New(days, startHour)
+	service, err := New(days, int64(startHour))
 	if err != nil {
 		g.Log().Fatal(ctx, "create esc service failed", "error", err)
 	}
